Donate Page Refactored and Wrapped
