+++ frontend/src/components/Donate/Donate.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import MakeDonation from './MakeDonation'
+import './style.css'
+const Donate = () => (
+  <div className="container pageDiv">
+    <h1>Donate</h1>
+    <MakeDonation />
+  </div>
+);
+
+export default Donate;
