+++ frontend/src/components/Donate/style.css
@@ -1,32 +1,19 @@
-/**
- * The CSS shown here will not be introduced in the Quickstart guide, but shows
- * how you can use CSS to style your Element's container.
- */
-.StripeElement {
-  box-sizing: border-box;
-
-  height: 50px;
-  width: 600px;
-
-  padding: 10px 12px;
-
-  border: 1px solid transparent;
-  border-radius: 4px;
-  background-color: white;
-
-  box-shadow: 0 1px 3px 0 #e6ebf1;
-  -webkit-transition: box-shadow 150ms ease;
-  transition: box-shadow 150ms ease;
+.title {
+  line-height: 2.5;
+  font-size: 45px;
+  color: #024885;
 }
 
-.StripeElement--focus {
-  box-shadow: 0 1px 3px 0 #cfd7df;
+.pageDiv {
+  font-family: "Abel", sans-serif;
+  width: 100vw;
+  min-height: 60vh;
+  padding: 50px 90px;
+  background-color: rgb(221, 221, 221);
+  color: #122f5e;
 }
 
-.StripeElement--invalid {
-  border-color: #fa755a;
-}
-
-.StripeElement--webkit-autofill {
-  background-color: #fefde5 !important;
+.mainPage {
+  width: 100vw;
+  min-height: 100%;
 }
